test: collect Windows cancellation suite by platform
